raspberry-pi/requirements.txt
# Core dependencies for Raspberry Pi Fruit Sorting System

# Camera and Image Processing
# picamera2>=0.3.17  # Preferred for Raspberry Pi Camera
opencv-python>=4.8.0   # Fallback camera + image processing
Pillow>=10.1.0
numpy>=1.24.0

# Hardware Control
RPi.GPIO>=0.7.1
gpiozero>=1.6.2

# Network Communication
pika>=1.3.2
requests>=2.31.0

# Web Framework
Flask>=3.0.0

# Configuration
python-dotenv>=1.0.0

# Logging and Utilities
loguru>=0.7.0
PyYAML>=6.0

# AI/ML (Optional - will install if available)
# ultralytics>=8.0.0
# tflite-runtime>=2.13.0

# Note: picamera2 may not install via pip on all systems
# If installation fails, the system will fallback to OpenCV camera
